Added Moving Animation

+++ ElementalityMagicWar/Assets/Source/Scripts/Health.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Health : MonoBehaviour
 {
     [SerializeField] private int maxHealth = 100;
-    [SerializeField] private int currentHealth;
+    [SerializeField] private float currentHealth;
 
     private void Start()
     {
@@ -16,7 +16,7 @@ public class Health : MonoBehaviour
     }
 
 
-    public void TakeDamage(int damageAmount)
+    public void TakeDamage(float damageAmount)
     {
         currentHealth -= damageAmount;
         if (currentHealth <= 0)
